revise learning rate

+++ Training.Conv.py
@@ -5,7 +5,7 @@
 # Training The ConvNet for Tensor Caller
 #========================================================================================================
 
-from optparse import OptionParser
+import argparse
 from datetime import datetime
 import time
 import os 
@@ -139,13 +139,95 @@ def train():
 				print 'Step %d Test loss = %.3f (%.3f sec)' % (step, loss_value, duration)
 			last_loss = loss_value
 
+def continue_train(ModelCKPT):
+	"""Train TensorCaller for a number of steps."""
+	with tf.Graph().as_default():
+		print "Locating Data File"
+		TrainingData = gzip.open(FLAGS.TrainingData,'rb')
+		TestingData = gzip.open(FLAGS.TestingData,'rb')
+		data_sets_training = Window2Tensor.Data_Reader(TrainingData, batch_size=BATCH_SIZE)
+		data_sets_testing = Window2Tensor.Data_Reader(TestingData, batch_size=BATCH_SIZE) 
+		print "Training Data @%s; \nTesting Data @%s" % (os.path.abspath(FLAGS.TrainingData), os.path.abspath(FLAGS.TestingData))
+
+		# Get Tensors and labels for Training data.
+		#tensors, labels = Models.inputs(FLAGS.data_file)
+
+		global_step = tf.Variable(0, trainable=False, name='global_step')
+
+		# Build a Graph that computes the logits predictions from the
+		# inference model.
+		tensor_placeholder, labels_placeholder = placeholder_inputs(BATCH_SIZE)
+		convnets = Models.ConvNets()
+		logits = convnets.Inference(tensor_placeholder)
+
+		# Calculate loss.
+		loss = convnets.loss(logits, labels_placeholder)
+
+		# Build a Graph that trains the model with one batch of examples and
+		# updates the model parameters.
+		train_op = convnets.Train(loss, global_step)
+		summary = tf.summary.merge_all()
+
+		#init = tf.global_variables_initializer()
+		saver = tf.train.Saver()
+
+		sess = tf.Session()
+		saver.restore(sess, ModelCKPT)
+		#print global_step
+		summary_writer = tf.summary.FileWriter(log_dir, sess.graph)
+		#sess.run(init)
+		
+		last_loss = 100
+		for step in xrange(max_steps):
+			start_time = time.time()
+			feed_dict = fill_feed_dict(data_sets_training, tensor_placeholder, labels_placeholder)
+			
+			_, loss_value = sess.run([train_op, loss], feed_dict=feed_dict)
+			duration = time.time() - start_time
+			if step % 10 == 0:
+				print 'Step %d Training loss = %.3f (%.3f sec)' % (step, loss_value, duration)
+				summary_str = sess.run(summary, feed_dict = feed_dict)
+				summary_writer.add_summary(summary_str, step)
+				summary_writer.flush()
+
+			if (step + 1) % 100 == 0 or (step + 1) == max_steps:
+				#Save Model only if loss decreasing
+				if loss_value < last_loss:
+					checkpoint_file = os.path.join(log_dir, 'model.ckpt')
+					saver.save(sess, checkpoint_file, global_step = step)
+				
+				feed_dict = fill_feed_dict(data_sets_testing, tensor_placeholder, labels_placeholder)
+				loss_value = sess.run(loss, feed_dict=feed_dict)
+				print 'Step %d Test loss = %.3f (%.3f sec)' % (step, loss_value, duration)
+			last_loss = loss_value
+
+def GetOptions():
+	parser = argparse.ArgumentParser()
+	parser.add_argument("-c", "--continue" help="continue training from a checkpoint",
+                    type=str)
+	args = parser.parse_args()
+	if args.continue.lower() in ['y', 'yes', 't', 'true']:
+		return True
+
 def main(argv=None):  # pylint: disable=unused-argument
-	
-	if tf.gfile.Exists(FLAGS.train_dir):
-		tf.gfile.DeleteRecursively(FLAGS.train_dir)
-	tf.gfile.MakeDirs(FLAGS.train_dir)
+	Continue = GetOptions()
+
 	print 'TraingDir is:',FLAGS.train_dir
-	train()
+	if Continue:
+		ckptfile = FLAGS.checkpoint_dir+'/log/checkpoint'
+		ckpt = open(ckptfile,rb).readline().split(':').strip().strip('"')
+		print ckpt
+		exit()
+		continue_train(ckpt)
+	else:
+		cmd = raw_input("Start a New Training?(y/n):")
+		if cmd == 'y':
+			if tf.gfile.Exists(FLAGS.train_dir):
+				tf.gfile.DeleteRecursively(FLAGS.train_dir)
+				tf.gfile.MakeDirs(FLAGS.train_dir)
+			train()
+		else:
+			exit()
 
 
 if __name__ == '__main__':
